# speed up / Use assert to check for corrupt mem